Create map show view, dark theme
This commit updates the map show page to more closely match the layout
sketch. The map is scrollable, and the page is designed with the idea of
being mobile first.

A dark theme is also added to the body of the application layout so that
each page has a dark grey background and white text by default.