Consolidate Dockerfiles, allow hot code reloading in container
This commit updates the Dockerfiles and docker-compose files to allow
for a development version that has hot code reloading and runs in a
container.

A series controller test is now passing, but a future commit will needs
to fix the issue with using path helpers without going to example.com.