Update Dockerfile, compose, and deploy script
This commit moves the command to start the server from the Dockerfile to
the docker-compose files. This facilitates the addition of the `cmd`
service, which can be used to run one-off commands in the context of the
app, without using the app service.

The deploy script for deploying the production server is updated to use
the proper docker-compose file as well.

All servers now run on port 35001, and the docker-compose file is
responsible for mapping that to the proper port based on the
environment.