Update dev deployment process
This commit updates the dev deployment process to also include using
webpack-dev-server for hot reloading of JS.

Since I can't yet get webpack-dev-server to run in the background, this
commit runs the rails server in the background. As a consequence, logs
aren't added to STDOUT, and must be checked using `log/development.log`.