Reduce Dockerfile steps, update dev deploy, set up swarm deploy for prod
This commit refactors the Dockerfile so that there are less steps, which
will hopefully reduce build times.

A yarn installation bug is fixed by updating apt-get after adding the
repo but before running the install command so that the correct version
is installed.

An (insecure) secret key base is added to production using the
environment variables in the docker compose file. A future commit will
replace this, tracked in issue #58
https://github.com/manik1235/story_time/issues/58

The `swarm.yml` docker compose file is set to listen on port 80, and is
set to deploy to swarm.

The `deploy.sh` script is renamed `deploy-prod.sh`, and updated to
deploy to swarm.

This commit also updates the dev deployment process. The script name is
changed from `deploy-dev.sh` to `start-dev.sh` to better match it's
funciton. The `deploy-*` scripts should actually deploy the container,
while the `start-*` scripts should be run as a command in a Dockerfile
or docker-compose file.