Draw first set of r axis lines, refactor _drawDashedLines
This commit adds the first set of r axis lines. They will be properly
placed in a future commit.

This commit also refactors the `_drawDashedLines` method to accept and
array for `hexDiameter`. The first element is the x step, and the 2nd is
the y step. This will allow the method to handle horizontal lines as
well as vertical and angled lines.