Only construct GameMap if element exists, simplify GameMap constructor
In order to prevent Jasmine tests from failing when the JS is loaded,
the `view/maps/show` javascript is updated to only construct the GameMap
if the selector exists.

The constructor for the GameMap class is updated to store the selector
instead of the element. The element is moved to a getter instead.