Refactor object serialization
This commit refactors the way objects are serialized. The
`active_model_serializers` gem is added to help handle the process. The
`#to_json` method for the map model is added and returns the serialized
output, including axes. The data attribute on the HTML map element is
replaced with a single data attribute containing the map and all axes.

The JavaScript is updated to be more agnostic to the number of axes and
their names.

The GameMap class is also refactored a bit to abstract some of the data
it's using.